Did explicit type conversion

+++ source/1.Getting_Started/3.Shaders/2.Green_To_Black_Ex2.cpp
@@ -135,7 +135,7 @@ int main() {
         glClear(GL_COLOR_BUFFER_BIT);
 
         float timeValue = (float)glfwGetTime();
-        float greenValue = (sin(timeValue) / 2.0f) + 0.5f;
+        float greenValue = ((float)sin(timeValue) / 2.0f) + 0.5f;
         int vertexColorLocation =
                 glGetUniformLocation(shaderProgram, "ourColor");
         glUseProgram(shaderProgram);
